requirements.txt
streamlit>=1.30.0
notion-client==2.2.1
python-dotenv